organization step 1: reformatted file directory
