+++ frontend/src/Components/SearchBar.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import "./App.css";
+import "../App.css";
 import Button from "react-bootstrap/Button";
 import "bootstrap/dist/css/bootstrap.min.css";
 // import ButtonToolbar from "react-bootstrap/ButtonToolbar";
