+++ frontend/src/Pages/ErrorPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import NavBar from "../NavBar";
+import NavBar from "../Components/NavBar";
 import { useNavigate } from "react-router-dom";
 
 function ErrorPage(props) {
