+++ frontend/src/Pages/Profile.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
-import NavBar from "../NavBar";
+import NavBar from "../Components/NavBar";
 import { useNavigate, Link } from "react-router-dom";
-import MovieList from "../MovieList";
+import MovieList from "../Components/MovieList";
 import { useCookies } from "react-cookie";
 
 function Profile(props) {
